question 4.3

+++ src/Controller/GameController.php
@@ -6,6 +6,7 @@ use App\Entity\Game;
 use App\Entity\UserCharacters;
 use App\Form\GameType;
 use App\Repository\GameRepository;
+use App\Security\AppAccess;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -57,6 +58,7 @@ class GameController extends AbstractController
      */
     public function show(Game $game): Response
     {
+        $this->denyAccessUnlessGranted(AppAccess::GAME, $game);
         return $this->render('game/show.html.twig', [
             'game' => $game,
         ]);
@@ -67,6 +69,8 @@ class GameController extends AbstractController
      */
     public function edit(Request $request, Game $game): Response
     {
+        $this->denyAccessUnlessGranted(AppAccess::GAME, $game);
+
         $form = $this->createForm(GameType::class, $game);
         $form->handleRequest($request);
 
@@ -89,6 +93,8 @@ class GameController extends AbstractController
      */
     public function delete(Request $request, Game $game): Response
     {
+        $this->denyAccessUnlessGranted(AppAccess::GAME, $game);
+
         if ($this->isCsrfTokenValid('delete'.$game->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($game);
